The structure of the document was changed. Some tags were deleted.
<meta> tag is renamed as <metaInfo> tag.
Also, <row> tags were deleted because now xml2map method doesn't
need them. The method assumes that <cell> elements in the document
are described in the way that the corresponding cell aligns
from left to right and from top to bottom of the map.
It also assumes that width and height of the map were described
in <metaInfo> element. Therefore, the method can calculate the
coordinate of the cell without the row number being explicitly
described.